+++ 1.21/Common/src/main/java/fuzs/puzzleslib/impl/core/EventHandlerProvider.java
@@ -2,11 +2,11 @@ package fuzs.puzzleslib.impl.core;
 
 public interface EventHandlerProvider {
 
-    void registerHandlers();
+    void registerEventHandlers();
 
     static void tryRegister(Object o) {
         if (o instanceof EventHandlerProvider provider) {
-            provider.registerHandlers();
+            provider.registerEventHandlers();
         }
     }
 }