+++ 1.21/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.0.20-1.21] - 2024-08-09
+### Added
+- Add `AddBlockEntityTypeBlocksCallback`
+### Fixed
+- Fix `ModelEvents` mixing up different block state models when baking on NeoForge
+- Fix `ModelEvents` breaking when ModernFix's dynamic resources option is enabled
+
 ## [v21.0.19-1.21] - 2024-08-07
 ### Changed
 - Disable config file watcher after too many reloads